Moving face and eye trackers out of interface
