+++ interface/src/Application.cpp
@@ -128,6 +128,7 @@
 #include <QmlWebWindowClass.h>
 #include <Preferences.h>
 #include <display-plugins/CompositorHelper.h>
+#include <trackers/EyeTracker.h>
 
 
 #include "AudioClient.h"
@@ -136,7 +137,6 @@
 #include "avatar/ScriptAvatar.h"
 #include "CrashHandler.h"
 #include "devices/DdeFaceTracker.h"
-#include "devices/EyeTracker.h"
 #include "devices/Leapmotion.h"
 #include "DiscoverabilityManager.h"
 #include "GLCanvas.h"
