+++ libraries/trackers/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(TARGET_NAME trackers)
+setup_hifi_library()
+GroupSources("src")
+link_hifi_libraries(shared)
+
+target_bullet()
