+++ libraries/trackers/src/trackers/Logging.cpp
