+++ libraries/trackers/src/trackers/MotionTracker.cpp
